requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python==4.8.0
Flask==2.3.3
Pillow==10.0.0
gunicorn==21.2.0
python-dotenv==1.0.0 